refactor: 컴포넌트 도메인별 분리 및 HTML 문자열 반환 통일
- 컴포넌트를 도메인별로 재구성: cart, product, help, shared/layouts
- ShoppingCartItemElement → ShoppingCartItem 네이밍 개선
- 모든 컴포넌트 함수가 HTML 문자열 반환하도록 통일
- createElement 사용 완전 제거, innerHTML/insertAdjacentHTML 방식 적용
- 클린 아키텍처 및 관심사 분리 원칙 적용
- 87개 테스트 모두 통과하여 기능 무결성 보장
