+++ src/basic/domains/help/components/index.js
@@ -1,0 +1,3 @@
+export * from "./HelpContentPanel";
+export * from "./HelpModalBackdrop";
+export * from "./HelpModalToggleButton";
