+++ src/basic/domains/product/components/ProductSelectionPanel.js
