+++ src/basic/domains/product/components/StockWarningMessage.js
