+++ src/basic/domains/product/components/index.js
@@ -1,0 +1,3 @@
+export * from "./ProductDropdownSelect";
+export * from "./ProductSelectionPanel";
+export * from "./StockWarningMessage";
