--- src/basic/domains/ui/index.js
@@ -1,2 +1,0 @@
-// Re-export all UI components
-export * from "./ui.js";
