--- src/basic/domains/ui/ui.js
@@ -1,13 +1,0 @@
-// Re-export all UI components from individual files
-export { CartAddButton } from "./CartAddButton.js";
-export { CartItemsContainer } from "./CartItemsContainer.js";
-export { HelpContentPanel } from "./HelpContentPanel.js";
-export { HelpModalBackdrop } from "./HelpModalBackdrop.js";
-export { HelpModalToggleButton } from "./HelpModalToggleButton.js";
-export { MainLayoutGrid } from "./MainLayoutGrid.js";
-export { OrderSummaryColumn } from "./OrderSummaryColumn.js";
-export { ProductDropdownSelect } from "./ProductDropdownSelect.js";
-export { ProductSelectionPanel } from "./ProductSelectionPanel.js";
-export { ShoppingAreaColumn } from "./ShoppingAreaColumn.js";
-export { ShoppingCartItemElement } from "./ShoppingCartItemElement.js";
-export { StockWarningMessage } from "./StockWarningMessage.js";
