+++ src/basic/shared/components/layouts/index.js
@@ -1,1 +1,3 @@
 export * from "./Header";
+export * from "./MainLayoutGrid"; 
+export * from "./ShoppingAreaColumn";
